requirements.txt
# CareVerify - Python Dependencies
# Production requirements

# ─── Web Framework ─────────────────────────────────────────
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0
gevent==23.9.1

# ─── Supabase ──────────────────────────────────────────────
supabase==2.3.4
PyJWT==2.8.0

# ─── Async Task Queue ──────────────────────────────────────
celery==5.3.6
redis==5.0.1
flower==2.0.1
kombu==5.3.4

# ─── ML / AI ───────────────────────────────────────────────
numpy==1.26.3
scikit-learn==1.4.0
xgboost==2.0.3
joblib==1.3.2

# ─── Deep Learning (optional, auto-skipped if GPU absent) ──
# tensorflow==2.15.0  # uncomment for autoencoder support

# ─── NLP ───────────────────────────────────────────────────
spacy==3.7.2
# Run after install: python -m spacy download en_core_web_sm

# ─── OCR ───────────────────────────────────────────────────
pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.2.0
PyMuPDF==1.23.8       # PDF → image conversion
pdfplumber==0.10.3    # Fallback PDF text extraction

# ─── Utilities ─────────────────────────────────────────────
python-dotenv==1.0.0
pydantic==2.5.3
marshmallow==3.20.2

# ─── Security ──────────────────────────────────────────────
cryptography==42.0.1
bleach==6.1.0

# ─── Observability ─────────────────────────────────────────
structlog==24.1.0
sentry-sdk[flask]==1.40.4
prometheus-flask-exporter==0.23.1